Add NamedResultSet type, adress review issues
